implementação da classe getById

+++ src/main/java/org/group2/core/repository/ProfessorRepository.java
@@ -9,6 +9,6 @@ public interface ProfessorRepository {
     void save(Professor professor);
     void delete(Long id);
     void update(Professor professor, Long id);
-    Professor getById(Long id);
+    Professor getById(int id);
     List<Professor> getAll();
 }
